refactor(code-tab): extract BrowseFileView; refresh stale diff-schema comment
- Move the inline `BrowseFileView` helper out of `CodeTab.tsx` into its
  own `right-panel/BrowseFileView.tsx`, per the "one component per file"
  rule. CodeTab stays a layout shell, BrowseFileView owns the
  fetch-then-render concern.
- Update `GitDiffOutputSchema`'s comment to point at `@pierre/diffs`'s
  `parsePatchFiles` (the current consumer), not the retired
  `@git-diff-view/core` parser.
