+++ packages/client/src/right-panel/CodeTab.tsx
@@ -36,7 +36,7 @@ import {
 } from "../ui/Icons";
 import PierreFileTree, { toGitStatusEntries } from "../ui/PierreFileTree";
 import PierreDiffView from "../ui/PierreDiffView";
-import PierreFileView from "../ui/PierreFileView";
+import BrowseFileView from "./BrowseFileView";
 import { COMPACT_ICON_BUTTON_CLASS } from "../ui/chromeSpacing";
 
 const EMPTY_STATE: Record<GitDiffMode, string> = {
@@ -349,45 +349,6 @@ const CodeTab: Component<{ meta: TerminalMetadata | null }> = (props) => {
         </Switch>
       </div>
     </Show>
-  );
-};
-
-/** File content viewer for browse mode. Reads the file via RPC and hands
- *  the contents to Pierre's `File` renderer for shiki-powered highlighting. */
-const BrowseFileView: Component<{
-  repoPath: string;
-  filePath: string;
-  theme: "light" | "dark";
-}> = (props) => {
-  const [fileContent] = createResource(
-    () => ({ repoPath: props.repoPath, filePath: props.filePath }),
-    (input) => client.fs.readFile(input),
-  );
-
-  return (
-    <Switch fallback={<div class="px-2 py-1 text-fg-3/50">Loading…</div>}>
-      <Match when={fileContent.error}>
-        <div class="px-2 py-1 text-danger">
-          Error: {(fileContent.error as Error).message}
-        </div>
-      </Match>
-      <Match when={fileContent()}>
-        {(fc) => (
-          <>
-            <Show when={fc().truncated}>
-              <div class="px-2 py-1 text-warning text-[10px] border-b border-edge bg-surface-1/30">
-                File truncated (exceeds 1 MB)
-              </div>
-            </Show>
-            <PierreFileView
-              name={props.filePath}
-              contents={fc().content}
-              theme={props.theme}
-            />
-          </>
-        )}
-      </Match>
-    </Switch>
   );
 };
